WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/019.html

<!doctype html>
<title>WebSockets: removeEventListener</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':80/ws/echo');
  ws.close();
  var events = ['open', 'message', 'error', 'close'];
  var got = [];
  var event;
  function addThis(e) {
    got.push(e.type);
  }
  for (var i = 0; i < events.length; ++i) {
    ws.addEventListener(events[i], addThis, false);
    ws.removeEventListener(events[i], addThis, false);
    event = document.createEvent('Event');
    event.initEvent(events[i], false, false);
    ws.dispatchEvent(event);
  }
  assert_equals(got.length, 0);
  if (got.length) {
    debug('Got: '+got);
  }
  end();
});
</script>
