LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/events/019-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: WebSocket is closed before the connection is established.
CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: WebSocket is closed before the connection is established.
CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: WebSocket is closed before the connection is established.
CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: WebSocket is closed before the connection is established.

PASS WebSockets: removeEventListener
PASS WebSockets: removeEventListener 1
PASS WebSockets: removeEventListener 2
PASS WebSockets: removeEventListener 3
